requirements.txt
pandas
requests
beautifulsoup4
python-dotenv
pyarrow 
